+++ src/pipecat/transports/local/audio.py
@@ -61,6 +61,8 @@ class LocalAudioInputTransport(BaseInputTransport):
         )
         self._in_stream.start_stream()
 
+        await self.set_transport_ready(frame)
+
     async def cleanup(self):
         await super().cleanup()
         if self._in_stream:
@@ -110,6 +112,8 @@ class LocalAudioOutputTransport(BaseOutputTransport):
             output_device_index=self._params.output_device_index,
         )
         self._out_stream.start_stream()
+
+        await self.set_transport_ready(frame)
 
     async def cleanup(self):
         await super().cleanup()