+++ src/pipecat/transports/network/small_webrtc.py
@@ -395,6 +395,7 @@ class SmallWebRTCInputTransport(BaseInputTransport):
             self._receive_audio_task = self.create_task(self._receive_audio())
         if not self._receive_video_task and self._params.video_in_enabled:
             self._receive_video_task = self.create_task(self._receive_video())
+        await self.set_transport_ready(frame)
 
     async def _stop_tasks(self):
         if self._receive_audio_task:
@@ -487,6 +488,7 @@ class SmallWebRTCOutputTransport(BaseOutputTransport):
         await super().start(frame)
         await self._client.setup(self._params, frame)
         await self._client.connect()
+        await self.set_transport_ready(frame)
 
     async def stop(self, frame: EndFrame):
         await super().stop(frame)