+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed an issue that could cause data to be sent to the transports when they
+  were still not ready.
+
 - Remove custom audio tracks from `DailyTransport` before leaving.
 
 ## [0.0.66] - 2025-05-02